Disable static allocation in the cosmos. Fix related issues: - In practice this decreases our allocation limits to 65k entities of each type, should be enough for the time being 	- we can later increase this limit by using unsigned instead of unsigned short for cosmic ids - Critical pool bug - max size was wrongly calculated under dynamic allocation, the max_size was converted to size_type which would give a far less number - Compilation errors due to some unit tests being enabled - Don't use linear cache in wandering pixels (why would we use it only for wandering pixels?)

+++ src/application/setups/editor/commands/create_layer_command.h
@@ -24,7 +24,7 @@ struct create_layer_command : allocating_command<editor_layer_id> {
 	}
 };
 
-using editor_layer_pool_size_type = unsigned short;
+using editor_layer_pool_size_type = unsigned;
 using layer_undo_free_input = augs::pool_undo_free_input<editor_layer_pool_size_type, editor_layer>;
 
 struct delete_layers_command {
