+++ src/application/setups/editor/resources/editor_resource_id.h
@@ -4,7 +4,7 @@
 #include "augs/misc/pool/pooled_object_id.h"
 #include "application/setups/editor/resources/all_editor_resource_types_declaration.h"
 
-using editor_resource_pool_size_type = unsigned short;
+using editor_resource_pool_size_type = unsigned;
 using editor_resource_type_id = type_in_list_id<all_editor_resource_types>;
 using editor_resource_pool_id = augs::pooled_object_id<editor_resource_pool_size_type>;
